a.m. & p.m.

+++ 112_lab1_7.c
@@ -1,37 +1,37 @@
 #include<stdio.h>
 int main()
 {
-    float time24,time12;
-    scanf("%f" ,&time24);
+    double time24,time12;
+    scanf("%lf" ,&time24);
     if(time24==12.00)
     {
-        printf("%.2f p.m." ,time24);
+        printf("%.2lf p.m." ,time24);
     }
     else if(time24==0.00)
     {
-        printf("%.2f a.m." ,time24);
+        printf("%.2lf a.m." ,time24);
     }
     else if(time24>12.00&&time24<24.00)
     {
         time12=time24-12;
         if(time24>0.00&&time24<12.00)
         {
-            printf("%.2f a.m." ,time12);
+            printf("%.2lf a.m." ,time12);
         }
         else if(time24>12.00&&time24<24.00)
         {
-            printf("%.2f p.m." ,time12);
+            printf("%.2lf p.m." ,time12);
         }
     }
     else
     {
         if(time24>0.00&&time24<12.00)
         {
-            printf("%.2f a.m." ,time24);
+            printf("%.2lf a.m." ,time24);
         }
         else if(time24>12.00&&time24<24.00)
         {
-            printf("%.2f p.m." ,time24);
+            printf("%.2lf p.m." ,time24);
         }
     }
     return 0;
